# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/sessions": {
      "post": {
        "operationId": "SessionController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Session created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Session name already exists"
          }
        },
        "summary": "Create a new WhatsApp session",
        "tags": [
          "sessions"
        ]
      },
      "get": {
        "operationId": "SessionController_findAll",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Max sessions to return (1-1000, default 1000)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of sessions to skip (for paging)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of sessions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List all sessions",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{id}": {
      "get": {
        "operationId": "SessionController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Get session by ID",
        "tags": [
          "sessions"
        ]
      },
      "delete": {
        "operationId": "SessionController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Session deleted"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Delete a session",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{id}/start": {
      "post": {
        "operationId": "SessionController_start",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Session already started"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Start a session and initialize WhatsApp connection",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{id}/stop": {
      "post": {
        "operationId": "SessionController_stop",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session stopped",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Stop a session and disconnect WhatsApp",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{id}/wake": {
      "post": {
        "description": "Reloads the engine for a session that was hibernated due to inactivity. Returns immediately with status INITIALIZING; poll GET /sessions/:id or listen to the session.status WebSocket event until it becomes READY.",
        "operationId": "SessionController_wake",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session resuming",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Resume a hibernated session (no QR scan required)",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{id}/logout": {
      "post": {
        "description": "Asks WhatsApp to remove this companion device, so it disappears from the account holder's Linked Devices list, then tears the session down. Unlike stop and delete — which only release the session locally and leave the device linked on the phone — reconnecting after a logout always requires a fresh QR scan or pairing code.",
        "operationId": "SessionController_logout",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session logged out",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Session is not started"
          },
          "404": {
            "description": "Session not found"
          },
          "502": {
            "description": "Session was stopped locally, but WhatsApp did not confirm the device unlink — retryable"
          }
        },
        "summary": "Log out of WhatsApp (unlinks this device) and stop the session",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{id}/force-kill": {
      "post": {
        "operationId": "SessionController_forceKill",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session force-killed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Force-kill a stuck session (SIGKILL its wedged engine, then tear it down)",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{id}/qr": {
      "get": {
        "operationId": "SessionController_getQRCode",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "QR code data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QRCodeResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "QR code not ready or session already authenticated"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Get QR code for session authentication",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{id}/pairing-code": {
      "post": {
        "operationId": "SessionController_requestPairingCode",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestPairingCodeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pairing code generated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairingCodeResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Session not started or already authenticated"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Request an 8-char pairing code to link via phone number (alternative to QR)",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{id}/groups": {
      "get": {
        "operationId": "SessionController_getGroups",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Max groups to return (1–1000, default 1000)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of groups to skip (for paging)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of groups the session is a member of"
          },
          "400": {
            "description": "Session not ready"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Get all groups for a session",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{id}/chats": {
      "get": {
        "operationId": "SessionController_getChats",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Max chats to return (1–1000, default 1000)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of chats to skip (for paging)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of active chats (most recent first)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChatSummaryDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Session not ready"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Get active chats for a session",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{id}/chats/read": {
      "post": {
        "operationId": "SessionController_markChatRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkChatReadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chat marked as read successfully"
          },
          "400": {
            "description": "Session not ready"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Mark a chat as read/seen",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{id}/chats/unread": {
      "post": {
        "operationId": "SessionController_markChatUnread",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkChatReadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chat marked as unread successfully"
          },
          "400": {
            "description": "Session not ready"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Mark a chat as unread",
        "tags": [
          "sessions"
        ]
      }
    },
    "/api/sessions/{id}/chats/delete": {
      "post": {
        "operationId": "SessionController_deleteChat",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteChatDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chat deleted successfully"
          },
          "400": {
            "description": "Session not ready"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Delete a chat from the chat list (e.g. a group you have left)",
        "tags": [
          "sessions"
        ]
      }
    },
...
    "/api/sessions/{sessionId}/groups": {
      "post": {
        "operationId": "GroupController_create",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Group created"
          }
        },
        "summary": "Create a new group",
        "tags": [
          "groups"
        ]
      }
    },
...
    "/api/stats/sessions/{sessionId}": {
      "get": {
        "operationId": "StatsController_getSessionStats",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Per-session statistics for the requested session."
          }
        },
        "summary": "Get statistics for a specific session",
        "tags": [
          "statistics"
        ]
      }
    },
...
  },
  "info": {
    "title": "OpenWA API",
    "description": "Open Source WhatsApp API Gateway - Free, Self-Hosted HTTP API",
    "version": "0.11.1",
    "contact": {
      "name": "OpenWA",
      "url": "https://github.com/rmyndharis/OpenWA",
      "email": "yudhi@rmyndharis.com"
    }
  },
...
  "servers": [
    {
      "url": "http://{host}:{port}",
      "description": "OpenWA instance",
      "variables": {
        "host": {
          "default": "localhost"
        },
        "port": {
          "default": "2785",
          "description": "PORT env var"
        }
      }
    }
  ],
  "components": {
    "securitySchemes": {
      "X-API-Key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "metrics-bearer": {
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "type": "http",
        "description": "METRICS_TOKEN for the GET /api/metrics scrape endpoint"
      }
    },
    "schemas": {
...
      "QRCodeResponseDto": {
        "type": "object",
        "properties": {
          "qrCode": {
            "type": "string",
            "description": "QR code as data URL",
            "example": "data:image/png;base64,..."
          },
          "status": {
            "type": "string",
            "enum": [
              "created",
              "initializing",
              "qr_ready",
              "authenticating",
              "ready",
              "disconnected",
              "action_required",
              "failed",
              "hibernated"
            ],
            "example": "qr_ready"
          }
        },
        "required": [
          "qrCode",
          "status"
        ]
      },
...
    }
  },
  "security": [
    {
      "X-API-Key": []
    }
  ]
}